Graphical changes + Cog Health
With the 12+ cog formula released, I consider this change urgent to let others know how the cogs work. The next update will likely deal with calculating cog defense. I do not know how this works, please submit an issue with feature request tag if you do know.
The other change is graphical, and regards gag buttons. I just thought it would look nice if it matched the TTR color even when nonorganic. I probably won't change to the new organic color because it's too similar to the regular gags and frankly I think it looks ugly on the app LOL.
Before I forget, there's also an indicator to show if gags are organic or not. This second row will very likely be where cog defense is added.