negative defense functional
negative defense at least appears to function how it does ingame currently, it does not yet appear on any of the indicators for level or history, but at minimum functions like other features in the calculator.
both types of defense can coexist, but i don't actually know what this behavior would look like ingame since it does not exist yet, but i think it may be possible at some point in the future.